requirements.txt
# Core dependencies for Vanguard-FEDformer
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities
PyYAML>=6.0
pathlib2>=2.3.0
tqdm>=4.64.0

# Data handling
h5py>=3.7.0
tables>=3.7.0

# Scientific computing
scipy>=1.9.0
statsmodels>=0.13.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.8.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Optional: GPU support
# cupy-cuda11x>=11.0.0  # Uncomment for CUDA 11.x support
# cupy-cuda12x>=12.0.0  # Uncomment for CUDA 12.x support

# Optional: Advanced features
# optuna>=3.0.0          # For hyperparameter optimization
# wandb>=0.13.0          # For experiment tracking
# mlflow>=1.30.0         # For model versioning